add unit tests and move functions due to loop dependencies
