+++ src/planner/parabolic/ParabolicUtil.cpp
@@ -6,19 +6,13 @@
 #include <dart/common/StlHelpers.hpp>
 #include <dart/dart.hpp>
 #include <aikido/common/Spline.hpp>
-#include <aikido/statespace/CartesianProduct.hpp>
 #include <aikido/statespace/GeodesicInterpolator.hpp>
-#include <aikido/statespace/Rn.hpp>
-#include <aikido/statespace/SO2.hpp>
 #include <aikido/trajectory/Interpolated.hpp>
 #include <aikido/trajectory/Spline.hpp>
 
 #include "DynamicPath.h"
 
 using Eigen::Vector2d;
-using aikido::statespace::CartesianProduct;
-using aikido::statespace::R;
-using aikido::statespace::SO2;
 using aikido::statespace::StateSpace;
 using dart::common::make_unique;
 
@@ -63,61 +57,6 @@ void evaluateAtTime(
   ParabolicRamp::Vector velocityVector;
   _path.Derivative(_t, velocityVector);
   _velocity = toEigen(velocityVector);
-}
-
-bool checkStateSpace(const statespace::StateSpace* _stateSpace)
-{
-  // TODO(JS): Generalize Rn<N> for arbitrary N.
-  if (dynamic_cast<const R<0>*>(_stateSpace) != nullptr)
-  {
-    return true;
-  }
-  else if (dynamic_cast<const R<1>*>(_stateSpace) != nullptr)
-  {
-    return true;
-  }
-  else if (dynamic_cast<const R<2>*>(_stateSpace) != nullptr)
-  {
-    return true;
-  }
-  else if (dynamic_cast<const R<3>*>(_stateSpace) != nullptr)
-  {
-    return true;
-  }
-  else if (dynamic_cast<const R<4>*>(_stateSpace) != nullptr)
-  {
-    return true;
-  }
-  else if (dynamic_cast<const R<5>*>(_stateSpace) != nullptr)
-  {
-    return true;
-  }
-  else if (dynamic_cast<const R<6>*>(_stateSpace) != nullptr)
-  {
-    return true;
-  }
-  else if (dynamic_cast<const R<Eigen::Dynamic>*>(_stateSpace) != nullptr)
-  {
-    return true;
-  }
-  else if (dynamic_cast<const SO2*>(_stateSpace) != nullptr)
-  {
-    return true;
-  }
-  else if (auto space = dynamic_cast<const CartesianProduct*>(_stateSpace))
-  {
-    for (std::size_t isubspace = 0; isubspace < space->getNumSubspaces();
-         ++isubspace)
-    {
-      if (!checkStateSpace(space->getSubspace<>(isubspace).get()))
-        return false;
-    }
-    return true;
-  }
-  else
-  {
-    return false;
-  }
 }
 
 std::unique_ptr<aikido::trajectory::Spline> convertToSpline(
