+++ tests/planner/kunzretimer/test_KunzRetimerPostProcessor.cpp
@@ -4,13 +4,14 @@
 #include <aikido/planner/parabolic/ParabolicTimer.hpp>
 #include <aikido/statespace/GeodesicInterpolator.hpp>
 #include <aikido/statespace/Rn.hpp>
+#include <aikido/trajectory/util.hpp>
 #include "eigen_tests.hpp"
 
 using Eigen::Vector2d;
 using aikido::trajectory::Interpolated;
 using aikido::statespace::GeodesicInterpolator;
 using aikido::statespace::R2;
-using aikido::planner::parabolic::convertToSpline;
+using aikido::trajectory::convertToSpline;
 using aikido::planner::kunzretimer::KunzRetimer;
 
 class KunzRetimerPostProcessorTests : public ::testing::Test
