+++ tests/planner/parabolic/test_TimePostProcessor.cpp
@@ -3,13 +3,14 @@
 #include <aikido/planner/parabolic/ParabolicTimer.hpp>
 #include <aikido/statespace/GeodesicInterpolator.hpp>
 #include <aikido/statespace/Rn.hpp>
+#include <aikido/trajectory/util.hpp>
 #include "eigen_tests.hpp"
 
 using Eigen::Vector2d;
 using aikido::trajectory::Interpolated;
 using aikido::statespace::GeodesicInterpolator;
 using aikido::statespace::R2;
-using aikido::planner::parabolic::convertToSpline;
+using aikido::trajectory::convertToSpline;
 using aikido::planner::parabolic::ParabolicTimer;
 
 class TimePostProcessorTests : public ::testing::Test
